testing and production mode added
